Remove Android support for out-of-process KeyStores
Support for out-of-process KeyStores was added to Chrome for Android
builds to support certain OEM-specific functionality. It worked by binding
to a service that implemented the RemoteAndroidKeyStoreInterface, and
using that for private key operations, rather than the default in-process
KeyStore.

However, support for this functionality in Chrome is being retired;
device manufacturers wishing to offer this functionality to Chrome
users can simply fall back to the default KeyStore implementation.

As such, remove the code and service descriptors; only local,
in-process KeyStores will be supported.

BUG=560614

Review URL: https://codereview.chromium.org/1474603004

Cr-Original-Commit-Position: refs/heads/master@{#362649}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e04b8f5353ab621c48b90bb46fee24fd0c0f4a29