Certificate Transparency: Ensure known logs are correctly parsed.
It is possible that the known logs are not correctly parsed in all clients,
leading to null logs being added to the list, crashing the client when
accessed.
This change would crash the client sooner so we can identify that is the
case.

BUG=493513

Review URL: https://codereview.chromium.org/1528403002

Cr-Original-Commit-Position: refs/heads/master@{#365610}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d15a868a4f72a485fce1f4e3c2d8a3b115369e53