Changes to AlternativeServices initialization to maintain MRU order.
If an entry from Preferences is not in the memory, it will be added to
the end of MRU list.

If an entry from Preferences is already in the memory, it will be
updated and its MRU order is maintained.

Entries in the memory that are not in Preferences, maintain their
recency.

BUG=568386
R=bnc@chromium.org, rch@chromium.org

Review URL: https://codereview.chromium.org/1531943003

Cr-Original-Commit-Position: refs/heads/master@{#366032}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 59fa432a5cd6569f5b5b9077d3c4c70d69cacb09