Teach URLRequest about initiator checks for First-Party-Only cookies.
This patch adds an 'initiator' field to 'URLRequest' in order to keep track
of the origin of the context which initiated a request. This allows us to
correctly perform the initiator requests specified in [1], which prevent
first-party-only cookies from being sent along with a request if that request
is "unsafe", and is initiated from a third-party origin (e.g. a form
submission to 'bank.com' from 'evil.com').

[1]: https://tools.ietf.org/html/draft-west-first-party-cookies-04#section-4.3

BUG=544114

Review URL: https://codereview.chromium.org/1411813003

Cr-Original-Commit-Position: refs/heads/master@{#369754}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 202534e3fa636aa1c9ce73c30dbbba854992488f