Certificate Transparency: Fetching consistency proofs.
This adds the ability for Chrome clients to fetch consistency proofs between two tree heads.
It will be used in case a Chrome client has an old STH stored for which it was not provided with a consistency proof to a new STH pushed in a component update.

BUG=506227

Review URL: https://codereview.chromium.org/1405293009

Cr-Original-Commit-Position: refs/heads/master@{#374865}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6851c0c562cb3f711f0fa03fff18a0750ac221e9