net: move GetHostName() function into network_interfaces.h
This patch moves it from net_util to network_interfaces because the
hostname is a name for one of the machine's addresses. There could be
multiple names based on the multiple addresses for the various
interfaces, so this seems like a reasonable place.

BUG=488531
TEST=net_unittests --gtest_filter=NetworkInterfacesTest.*
R=eroman@chromium.org
TBR=darin@chromium.org # for changes outside of net

Review URL: https://codereview.chromium.org/1696533002

Cr-Original-Commit-Position: refs/heads/master@{#375102}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e5f66e6128834895ee2a31206542d586a5b75d97