Evict non-secure cookies less agressively.
The current implementation of strict-secure cookies wipes all non-secure
cookies when an origin exceeds ~180. This is a bit agressive, and has
real impact on users.

This patch softens that to remove only the number of non-secure cookies
necessary to get under the ~150 cap. If a user has 150 secure cookies,
we'll still wipe all non-secure cookies, but that seems less likely to
impact users than the current behavior.

The patch is a bit more complicated than I expected due to interactions
with the Chrome-only "priority" feature we shipped back in 2013. In
short, we execute priority-driven removal of non-secure cookies first,
and only touch secure cookies if necessary.

BUG=581588
R=mmenke@chromium.org, jww@chromium.org

Review URL: https://codereview.chromium.org/1705753002

Cr-Original-Commit-Position: refs/heads/master@{#376204}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 162d27135f2ee44ae01341de055d1b827a930767