Don't decompose a der::Tag when processing it.
Decomposing the number and class is a little iffy (one can make a CHOICE of
SEQUENCE and [0]). Treating the constructed bit as part of the tag is, in
the general BER case, not quite right, but it is a valid simplification in DER.

This pattern makes it less likely that we check the number without the class,
etc.

BUG=none

Review URL: https://codereview.chromium.org/1710613002

Cr-Original-Commit-Position: refs/heads/master@{#376313}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fb4788dd358c2bbefbb18aec82544ad103201f12