net: change SetDisableNagle() function to return bool type
This patch changes SetDisableNagle() function in tcp_socket_util to
return a bool type because there is no good reason to use an int type
and 0 and 1 to mean false/true, as we are not in C land but the C++
side.

While at it, move AddrinfoGuard type and CloseSocket() function into
unnamed namespace and fix a wrong reference to SO_REUSEADDR in reuseport
path.

BUG=None
TEST=net_unittests
R=rch@chromium.org

Review URL: https://codereview.chromium.org/1721483002

Cr-Original-Commit-Position: refs/heads/master@{#376736}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9484959dc07badc86a08cbd503c7dc90c128f8cd