net: rename SetDisableNagle() function to SetTCPNoDelay()
This patch renames SetDisableNagle() function to SetTCPNoDelay()
in preparation for merging it with two other functions that have the
same functionality, SetTCPNoDelay() in
net/socket/tcp_socket_posix.cc and DisableNagle() in
net/socket/tcp_socket_win.cc.

In order to make the merging step easier, we push the logging of
function failure outside of it, so the client handles it, not the
function in the library side.

BUG=None
TEST=net_unittests
R=eroman@chromium.org

Review URL: https://codereview.chromium.org/1742503002

Cr-Original-Commit-Position: refs/heads/master@{#377850}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3f6bb117cd1e97a7c2c91ce95dfaee89e9078604