Make CookieStore no longer threadsafe.
In practice, it's only used on one thread, and it's one of the few net
classes that tries to be threadsafe.  This also removes a lot of
locking, which Chromium code generally tries to avoid doing.

Other than removing thread-safety, and invoking FlushStore's
callback on the network thread, the API is unchanged.

BUG=46185

Review URL: https://codereview.chromium.org/1698693002

Cr-Original-Commit-Position: refs/heads/master@{#378505}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: be0910d6b03b3433448d25587ac57035d227b0d9