net: remove CanonicalCookie ctor that takes two parameters
The problem of using the constructor is that with it there
is no way to return an error if bad arguments are passed in.

That can be done by using the static method Create, that
does some validation and returns null if an attribute value
is invalid.

BUG=57061
TEST=unit_tests, net_unittests
R=mmenke@chromium.org
TBR=bauerb@chromium.org,msw@chromium.org

Review URL: https://codereview.chromium.org/1833483002

Cr-Original-Commit-Position: refs/heads/master@{#383069}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b1eaeaefa02ccc5a8fe79b2776a6aa34b14a24e0