Remove outdated cert verification tests
Remove CertVerifyProcTest.GlobalSignR3EVTest, as it hasn't been enabled
in years, and the underlying issue was not something that warrants a
brittle test like that.

Also remove the test that cert verification succeeds when revocation is
disabled, as the vast majority of the CertVerifyProc tests already test
that, and this test was disabled.

BUG=488084

Review URL: https://codereview.chromium.org/1831193003

Cr-Original-Commit-Position: refs/heads/master@{#383237}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aab2b180e98bfec07845113bea5e35ee29c1aabe