Log urls and headers in BIDIRECTIONAL_STREAM events
BIDIRECTIONAL_STREAM events do not have urls and headers in NetLog. This makes
debugging hard. Headers are currently logged in HTTP2_SESSION and QUIC_SESSION,
but the session events can have a bunch of requests going on, and it is hard to
associate headers with each bidirectional stream.

This CL logs urls and the headers sent and received in BIDIRECTIONAL_STREAM
events.

BUG=607613

Review-Url: https://codereview.chromium.org/1932853002
Cr-Original-Commit-Position: refs/heads/master@{#390518}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d4c21ab215ec191359bd8bb3f85f1249e219845c