Unify HTTP/2 and SPDY stream schedulers under common WriteScheduler interface.
This is to make it possible for SpdyDispatcher to maintain a single set of
codepaths for dealing with stream scheduling, with differences in scheduling
schemes hidden behind the WriteScheduler interface.  The actual wiring of
WriteScheduler into SpdyDispatcher will be in a follow-on CL.

This CL lands server change 115262424 by mpw.

BUG=488484

Review-Url: https://codereview.chromium.org/1900403004
Cr-Original-Commit-Position: refs/heads/master@{#393758}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 28130cb63807bc4bc81a176ada9d22ab3f5daa9e