service worker: Wait for request body blobs to finish construction
Blob construction recently became async and we have to handle that.

Note to sheriff: there's a good chance the crash test added will be
flaky on debug builds. That's because of an existing crash
crbug.com/611317 which happens before this patch. In that event,
I'd rather we add the test to TestExpectations than revert the
patch.

BUG=608544

Review-Url: https://codereview.chromium.org/1967683003
Cr-Original-Commit-Position: refs/heads/master@{#394241}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8e6cf517c56051b81e32c05dd2dddb024c93ffe2