Support setReadTimeout in CronetHttpURLConnection
This CL adds support for setReadTimeout in
CronetHttpURLConnection. Because of late binding of sockets
to requests, we cannot support per-request connect timeout.
This CL adds an UnsupportedOperationException when
setConnectTimeout is invoked.

The implementation part of the CL is mostly taken from
a patch by rohitagr@. This CL also adds unit tests.

BUG=611851

Review-Url: https://codereview.chromium.org/1984723002
Cr-Original-Commit-Position: refs/heads/master@{#394541}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5d1f9896cfc04ffcd76f5995c5725891a0ed8882