Introduce CertVerifier::RequestParams
In anticipation of divorcing caching from multi-threading,
move the MultiThreadedCertVerifier's RequestParams up to
CertVerifier, as part of a public, stable interface for
handling verification request. This mirrors the notion of
HostResolver's RequestInfo.

While Verify() does not yet take a RequestParams (which
will come later), this provides a common interface to use
as a key for determining if two requests should be treated
the same (e.g. for purposes of caching or joining existing
requests).

BUG=612655

Review-Url: https://codereview.chromium.org/1987113002
Cr-Original-Commit-Position: refs/heads/master@{#394967}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e51d72bacc26ac549ee7de34b6a96916e89cc5a9