Estimate Throughput across all requests
NQE is notified when the start of request finishes. NQE
keeps track of number of active requests that go over the
network, and those that do not go over the network.

A time-window may be started when the request is started
or when a request completes.

A throughput observation is taken over a time-window if
(i) traffic during that window was large enough
(more than a certain number of bytes); (ii) there was at
least one active request that went over the network at all
times during that window; and, (iii) there was no request
that was served by localhost or cache during that window.

Design doc: https://docs.google.com/document/d/1JYUDUfOcRStrjRzTrUeJ8VHT4RO4SLXq4Ha2pHsh9GI/edit#

This CL does not implement the check for hanging GETs which
will come in a subsequent CL.

BUG=502430

Review-Url: https://codereview.chromium.org/1916063002
Cr-Original-Commit-Position: refs/heads/master@{#395112}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 80a521639111557e14fee8db957a60fe4dc0a991