Record the number of freshness periods since a stale HttpCache entry was last used
This metrics records the number of freshness periods that have elapsed since a stale entry was last used. This will help us to determine the right time to prefresh resources (e.g., prefresh on cache hit if the entry's age is > 50% of freshness).

BUG=614355

Review-Url: https://codereview.chromium.org/2005753004
Cr-Original-Commit-Position: refs/heads/master@{#395885}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 91ae3602104125b4afca4f76d2b2841fda80718f