Move FROM_HERE out of InFlightBackendIO::PostOperation.
Currently, all types of BackendIO operations appear the same way in
chrome://tracing because they are all annotated with the FROM_HERE
macro located in InFlightBackendIO::PostOperation. This CL moves the
FROM_HERE macro to callers of InFlightBackendIO::PostOperation. This
allows the various types of BackendIO operation to be differentiated
in chrome://tracing.

BUG=

Review-Url: https://codereview.chromium.org/2013523004
Cr-Original-Commit-Position: refs/heads/master@{#396188}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 10981ba5e60e74ae3aeae2c4c880d114ebcbf23c