Fill in the current host name in the Mojo host resolver service when requesting our own IP address.
The V8 proxy resolver uses GetHostName() as the host in the
implementation of myIpAddress(). However, that may not return the
expected hostname when running inside the sandbox. Instead, use the
request's "is_my_ip_address" field to fill in the correct host name in
the browser before forwarding the resolve request to the DNS resolver.

BUG=615084

Review-Url: https://codereview.chromium.org/2011293002
Cr-Original-Commit-Position: refs/heads/master@{#396424}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3a6cd21dab6a285179b02a05e980091c86545d66