Disable AltSvc from insecure origin.
* Do not parse Alt-Svc response on insecure connections, add regression test.
* Do not follow in-memory Alt-Svc on insecure connections, add regression test.

This CL has a shortcut: |enable_alt_svc_for_insecure_origin|, which is false in
production but true in tests (except for the two regression tests).  Adapting
the existing tests is a lot of work, that will come in one or more follow-up
CLs.

BUG=615497

Review-Url: https://codereview.chromium.org/2026863002
Cr-Original-Commit-Position: refs/heads/master@{#397097}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e3dd56f4d98843611954b84fa28c1e884594fd05